refactor: Rank가 Cash 사용하도록 변경

+++ src/main/java/domain/LottoResult.java
@@ -23,7 +23,7 @@ public class LottoResult {
     }
 
     public BigDecimal winningRate() {
-        return countByRank.winningMoney().divide(cash.toBigDecimal(), 2, RoundingMode.HALF_UP);
+        return countByRank.winningMoney().toBigDecimal().divide(cash.toBigDecimal(), 2, RoundingMode.HALF_UP);
     }
 
     public int count(Rank rank) {
